+++ spec/tests/Project.spec.mjs
@@ -1,4 +1,4 @@
-import HashMap from "../../modules/HashMap.mjs";
+import HashMap from "../../src/HashMap.mjs";
 
 const test = HashMap();
 
@@ -24,14 +24,16 @@ describe("Project Odin Tests: ", function () {
     expect(test.get("carrot")).toBe("crunchy");
   });
   it("Overwrite lion's value with cat", function () {
+    console.log(test)
     test.set("lion", "cat");
     expect(test.get("lion")).toBe("cat");
-  });})
+  });
 
-//   it("Expand the map by one bucket", function () {
-//     expect(test.bucketSize()).toBe(15);
-//     test.set("moon", "silver");
-//     test.set("sun", "orange");
-//     expect(test.bucketSize()).toBe(16);
-//   });
-// });
+  // it("Expand the map by one bucket", function () {
+  //   expect(test.bucketSize()).toBe(15);
+  //   test.set("moon", "silver");
+  //   test.set("sun", "orange");
+  //   console.log(test)
+  //   expect(test.bucketSize()).toBe(16);
+  // });
+});
