+++ src/Get.mjs
@@ -1,20 +1,21 @@
-import Hash from "./Hash.mjs";
 import GetBucket from "./GetBucket.mjs";
+import Hash from "./Hash.mjs";
 
-export default function Remove(key, arr) {
+export default function Get(key, arr) {
+  // console.log(`${key} : ${arr}`);
+
   let myBucket;
   if (Array.isArray(arr)) {
     const hashed = Hash(key);
     myBucket = GetBucket(arr, arr.length, hashed);
   } else myBucket = arr;
-
-  if (Object.hasOwn(myBucket, key)) {
-    return delete myBucket[key];
-  } else if (Object.keys(myBucket).length == 0) {
+  if (key in myBucket) {
+    return myBucket[key];
+  } else if (myBucket === {}) {
     return false;
   } else if (myBucket["next"] == null) {
     return false;
   } else {
-    return Remove(key, myBucket["next"]);
+    return Get(key, myBucket["next"]);
   }
 }
