+++ src/Hash.mjs
