+++ src/Keys.mjs
@@ -2,12 +2,12 @@ export default function Keys(arr) {
   let keyList = [];
   if (Array.isArray(arr)) {
     arr.forEach((object) => {
-      if (Object.keys(object).length == 0) {
-        return;
+      if (Object.keys(object).length === 0) {
+
       } else {
         const objKeys = Object.keys(object);
         objKeys.forEach((element) => {
-          if (element != "next") {
+          if (element !== "next") {
             keyList.push(element);
           }
         });
