+++ src/LinkedList/Node.mjs
