+++ src/Remove.mjs
@@ -1,21 +1,20 @@
-import GetBucket from "./GetBucket.mjs";
 import Hash from "./Hash.mjs";
+import GetBucket from "./GetBucket.mjs";
 
-export default function Get(key, arr) {
-  // console.log(`${key} : ${arr}`);
-
+export default function Remove(key, arr) {
   let myBucket;
   if (Array.isArray(arr)) {
     const hashed = Hash(key);
     myBucket = GetBucket(arr, arr.length, hashed);
   } else myBucket = arr;
-  if (key in myBucket) {
-    return myBucket[key];
-  } else if (myBucket == {}) {
+
+  if (Object.hasOwn(myBucket, key)) {
+    return delete myBucket[key];
+  } else if (Object.keys(myBucket).length === 0) {
     return false;
   } else if (myBucket["next"] == null) {
     return false;
   } else {
-    return Get(key, myBucket["next"]);
+    return Remove(key, myBucket["next"]);
   }
 }
